Update: use secondary internal represenation that's more flat
Moving to a representation that's just an array of samples.
Each samples holds it's collection time.
It's not very visually inspectable,
but it's easily transformed into a time based one
(as we have in the original log).
Or it can be filtered/mapped/sorted into anything else easily

All the code should be migrated to this format eventually.
It makes everything much much easier!